Complete coroutine implementation
