+++ groups/ntc/ntc.t.cpp
@@ -411,7 +411,6 @@
 // ntcf
 
 // #include <ntcf_api.t.cpp>
-#include <ntcf_concurrent.t.cpp>
 #include <ntcf_system.t.cpp>
 #include <ntcf_test.t.cpp>
