+++ groups/ntc/ntcf/package/ntcf.mem
@@ -1,4 +1,3 @@
 ntcf_api
 ntcf_system
-ntcf_concurrent
 ntcf_test
